Fix connecting to a crashed runner while booting
This change adds a second field to the response from a slave node to
the client handler that indicates the state of the slave node. This
allows the server to detect connection to a runner that crashed. When
zeus is still booting, this should send a backtrace to the
client (like it does when Zeus has fully booted).