Make statuschart less verbose when logging to a non-TTY
On a non-TTY, use fewer lines to say more: Print the process tree as a
nested list on one line, with both colors and status abbreviations (to
allow copying & pasting of log files). Also, print all runnable
commands on one line and all crashed commands on another.

This cuts down the vertical space required to log state of Zeus when
booting to less than a screenful on a large process tree, and allows
easier diagnosis of boot problems in, e.g., a chat room.