Pass a pipe for args from zeus client to command
Instead of passing the command arguments from the zeus client to the
zeus master, and from there to the command process over their respective
sockets, create a streaming socket in the zeus client, pass that socket
over to the zeus master, have the zeus master pass it on to the command
process, and read the zero-separated arguments directly in the command.

This has a two advantages:

One, it supports unbounded numbers of arguments on Linux and OS
X (previous solutions only worked either on one or the other).

Two, it extracts all command argument handling from the zeus master -
all the master now does is pass a count and an FD from one pipe to
another.