Report features to master before reporting errors
As soon as we report an error to the master, it kills our process.
This in turn prevents us from reporting the features loaded before
the error and getting restarted when those change.
https://github.com/burke/zeus/blob/0cd1c4fad7c214a4a842665922803b7e2bac3347/go/processtree/slavenode.go#L223-L226

Instead, we want to report features before reporting the error but ensure
that we eventually report the error even if feature reporting fails.