Fix debouncing logic
The file monitor fileChangeDelay parameter should act to debounce
change events. It should only report an event when it has not seen a
new change for at least fileChangeDelay. Currently Zeus implements
throttling as opposed to debouncing -- it sends an event at most every
fileChangeDelay but does not wait for the event stream to quiesce
before sending the first event. This can result in reloading code in
the midst of an operation such as an rsync or git change.